feat: 新增basic auth授权方式
