+++ git-checkout-core/src/main/kotlin/com/tencent/bk/devops/git/core/service/GitSourceProvider.kt
@@ -42,7 +42,7 @@ import com.tencent.bk.devops.git.core.service.handler.GitSubmodulesHandler
 import com.tencent.bk.devops.git.core.service.handler.HandlerExecutionChain
 import com.tencent.bk.devops.git.core.service.handler.InitRepoHandler
 import com.tencent.bk.devops.git.core.service.handler.PrepareWorkspaceHandler
-import com.tencent.bk.devops.git.core.service.helper.GitAuthHelperFactory
+import com.tencent.bk.devops.git.core.service.helper.auth.GitAuthHelperFactory
 import com.tencent.bk.devops.git.core.util.EnvHelper
 import com.tencent.bk.devops.git.core.util.GitUtil
 import org.slf4j.LoggerFactory
