+++ git-checkout-core/src/main/kotlin/com/tencent/bk/devops/git/core/service/helper/auth/CredentialCheckoutAuthHelper.kt
@@ -25,16 +25,18 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.tencent.bk.devops.git.core.service.helper
+package com.tencent.bk.devops.git.core.service.helper.auth
 
 import com.tencent.bk.devops.git.core.constant.ContextConstants
 import com.tencent.bk.devops.git.core.constant.GitConstants
 import com.tencent.bk.devops.git.core.constant.GitConstants.BK_CI_BUILD_JOB_ID
+import com.tencent.bk.devops.git.core.constant.GitConstants.BK_CI_BUILD_TASK_ID
 import com.tencent.bk.devops.git.core.constant.GitConstants.CREDENTIAL_JAR_PATH
 import com.tencent.bk.devops.git.core.constant.GitConstants.CREDENTIAL_JAVA_PATH
 import com.tencent.bk.devops.git.core.constant.GitConstants.GIT_CREDENTIAL_COMPATIBLEHOST
 import com.tencent.bk.devops.git.core.constant.GitConstants.GIT_CREDENTIAL_HELPER
 import com.tencent.bk.devops.git.core.constant.GitConstants.GIT_CREDENTIAL_HELPER_VALUE_REGEX
+import com.tencent.bk.devops.git.core.constant.GitConstants.GIT_CREDENTIAL_TASKID
 import com.tencent.bk.devops.git.core.constant.GitConstants.GIT_REPO_PATH
 import com.tencent.bk.devops.git.core.enums.AuthHelperType
 import com.tencent.bk.devops.git.core.enums.GitConfigScope
@@ -42,6 +44,7 @@ import com.tencent.bk.devops.git.core.enums.GitProtocolEnum
 import com.tencent.bk.devops.git.core.pojo.CredentialArguments
 import com.tencent.bk.devops.git.core.pojo.GitSourceSettings
 import com.tencent.bk.devops.git.core.service.GitCommandManager
+import com.tencent.bk.devops.git.core.service.helper.VersionHelper
 import com.tencent.bk.devops.git.core.util.CommandUtil
 import com.tencent.bk.devops.git.core.util.EnvHelper
 import org.apache.commons.codec.digest.DigestUtils
@@ -56,13 +59,13 @@ import java.nio.file.Paths
  * 使用自定义git-checkout-credential凭证
  */
 @Suppress("ALL")
-class CredentialAuthHelper(
+class CredentialCheckoutAuthHelper(
     private val git: GitCommandManager,
     private val settings: GitSourceSettings
-) : HttpGitAuthHelper(git = git, settings = settings) {
+) : CredentialAuthHelper(git = git, settings = settings) {
 
     companion object {
-        private val logger = LoggerFactory.getLogger(CredentialAuthHelper::class.java)
+        private val logger = LoggerFactory.getLogger(CredentialCheckoutAuthHelper::class.java)
     }
 
     private val credentialVersion = VersionHelper.getCredentialVersion()
@@ -98,6 +101,15 @@ class CredentialAuthHelper(
         )
         EnvHelper.putContext(GitConstants.GIT_CREDENTIAL_AUTH_HELPER, AuthHelperType.CUSTOM_CREDENTIAL.name)
         eraseOauth2Credential()
+        git.tryConfigUnset(configKey = GIT_CREDENTIAL_HELPER)
+        if (git.isAtLeastVersion(GitConstants.SUPPORT_EMPTY_CRED_HELPER_GIT_VERSION)) {
+            git.tryDisableOtherGitHelpers(configScope = GitConfigScope.LOCAL)
+        }
+        git.config(configKey = GIT_CREDENTIAL_TASKID, configValue = settings.pipelineTaskId)
+        git.configAdd(
+            configKey = GIT_CREDENTIAL_HELPER,
+            configValue = "!bash '$credentialShellPath'"
+        )
         install()
         store()
     }
@@ -223,6 +235,7 @@ class CredentialAuthHelper(
             return
         }
         // 清理构建机上凭证
+        val taskId = git.tryConfigGet(configKey = GIT_CREDENTIAL_TASKID)
         if (File(credentialJarPath).exists()) {
             with(URL(settings.repositoryUrl).toURI()) {
                 CommandUtil.execute(
@@ -235,7 +248,8 @@ class CredentialAuthHelper(
                         "devopsErase"
                     ),
                     runtimeEnv = mapOf(
-                        GIT_REPO_PATH to settings.repositoryPath
+                        GIT_REPO_PATH to settings.repositoryPath,
+                        BK_CI_BUILD_TASK_ID to taskId
                     ),
                     inputStream = CredentialArguments(
                         protocol = scheme,
@@ -245,5 +259,11 @@ class CredentialAuthHelper(
                 )
             }
         }
+        git.tryConfigUnset(configKey = GIT_CREDENTIAL_TASKID)
+    }
+
+    override fun addSubmoduleCommand(commands: MutableList<String>) {
+        commands.add("git config credential.helper !bash '$credentialShellPath'")
+        commands.add("git config credential.taskId ${settings.pipelineTaskId}")
     }
 }
