+++ git-checkout-credential/src/main/kotlin/com/tencent/bk/devops/git/credential/helper/GitHelper.kt
@@ -29,25 +29,21 @@ package com.tencent.bk.devops.git.credential.helper
 
 import com.tencent.bk.devops.git.credential.ConfigScope
 import com.tencent.bk.devops.git.credential.Constants.GIT_REPO_PATH
-import java.io.File
-import java.io.InputStream
 import org.apache.commons.exec.CommandLine
 import org.apache.commons.exec.DefaultExecutor
 import org.apache.commons.exec.LogOutputStream
 import org.apache.commons.exec.PumpStreamHandler
 import org.apache.commons.exec.environment.EnvironmentUtils
+import java.io.File
+import java.io.InputStream
 
 object GitHelper {
 
     fun tryConfigGet(
         configKey: String,
-        configValueRegex: String? = null,
-        configScope: ConfigScope? = ConfigScope.LOCAL
+        configValueRegex: String? = null
     ): String? {
         val args = mutableListOf("config")
-        if (configScope != null) {
-            args.add(configScope.option)
-        }
         args.add("--get")
         args.add(configKey)
         if (!configValueRegex.isNullOrBlank()) {
